Modified Cobalt plugin to split it into two shared objects.
