+++ Cobalt/CMakeLists.txt
@@ -24,6 +24,7 @@ find_package(WPEFramework)
 project_version(1.0.0)
 
 set(MODULE_NAME ${NAMESPACE}${PROJECT_NAME})
+set(PLUGIN_COBALT_IMPLEMENTATION "${MODULE_NAME}Impl" CACHE STRING "library with cobalt implementation." )
 
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/cmake")
 
@@ -44,11 +45,14 @@ endif()
 
 add_library(${MODULE_NAME} SHARED 
     Cobalt.cpp
-    CobaltImplementation.cpp
     CobaltJsonRpc.cpp
     Module.cpp)
 
-target_compile_definitions(${MODULE_NAME} 
+add_library(${PLUGIN_COBALT_IMPLEMENTATION} SHARED
+    Module.cpp
+    CobaltImplementation.cpp)
+
+target_compile_definitions(${PLUGIN_COBALT_IMPLEMENTATION}
     PRIVATE 
         STARBOARD 
         STARBOARD_CONFIGURATION_INCLUDE="${WPEFRAMEWORK_PLUGIN_COBALT_STARBOARD_CONFIGURATION_INCLUDE}")
@@ -56,12 +60,18 @@ target_compile_definitions(${MODULE_NAME}
 target_link_libraries(${MODULE_NAME} 
     PRIVATE 
         ${NAMESPACE}Plugins::${NAMESPACE}Plugins
+        )
+
+target_link_libraries(${PLUGIN_COBALT_IMPLEMENTATION}
+    PRIVATE
+        ${NAMESPACE}Plugins::${NAMESPACE}Plugins
         ${COBALT_LIBRARIES}
         )
 
 # Library installation section
 string(TOLOWER ${NAMESPACE} STORAGENAME)
 install(TARGETS ${MODULE_NAME} DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/${STORAGENAME}/plugins)
+install(TARGETS ${PLUGIN_COBALT_IMPLEMENTATION} DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/${STORAGENAME}/plugins)
 
 set(PLUGIN_COBALT_AUTOSTART "false" CACHE STRING "Automatically start Cobalt plugin")
 set(PLUGIN_COBALT_MODE "Local" CACHE STRING "Controls if the plugin should run in its own process, in process or remote.")
