+++ Cobalt/Cobalt.config
@@ -3,6 +3,7 @@ set (preconditions Platform Graphics Internet)
 
 map()
     kv(mode ${PLUGIN_COBALT_MODE})
+    kv(locator lib${PLUGIN_COBALT_IMPLEMENTATION}.so)
     if(PLUGIN_COBALT_USER)
         kv(user ${PLUGIN_COBALT_USER})
     endif()
