+++ Cobalt/CobaltImplementation.cpp
@@ -624,53 +624,6 @@ SERVICE_REGISTRATION(CobaltImplementation, 1, 0);
 }
 /* namespace Plugin */
 
-namespace Cobalt {
-
-class MemoryObserverImpl: public Exchange::IMemory {
-private:
-    MemoryObserverImpl();
-    MemoryObserverImpl(const MemoryObserverImpl&);
-    MemoryObserverImpl& operator=(const MemoryObserverImpl&);
-
-    public:
-    MemoryObserverImpl(const RPC::IRemoteConnection* connection) :
-        _main(connection == nullptr ? Core::ProcessInfo().Id() : connection->RemoteId()) {
-    }
-    ~MemoryObserverImpl() {
-    }
-
-    public:
-    virtual uint64_t Resident() const {
-        return _main.Resident();
-    }
-    virtual uint64_t Allocated() const {
-        return _main.Allocated();
-    }
-    virtual uint64_t Shared() const {
-        return _main.Shared();
-    }
-    virtual uint8_t Processes() const {
-        return (IsOperational() ? 1 : 0);
-    }
-
-    virtual bool IsOperational() const {
-        return _main.IsActive();
-    }
-
-    BEGIN_INTERFACE_MAP (MemoryObserverImpl)INTERFACE_ENTRY (Exchange::IMemory)END_INTERFACE_MAP
-
-private:
-    Core::ProcessInfo _main;
-    };
-
-    Exchange::IMemory* MemoryObserver(const RPC::IRemoteConnection* connection) {
-        ASSERT(connection != nullptr);
-        Exchange::IMemory* result = Core::Service<MemoryObserverImpl>::Create<Exchange::IMemory>(connection);
-        return (result);
-    }
-}
-
-
 ENUM_CONVERSION_BEGIN(Plugin::CobaltImplementation::connection)
 
     { Plugin::CobaltImplementation::connection::CABLE,    _TXT("cable")    },
